Parser now requires type annotations.
There are going to be a number of steps involved in getting from here to
a fully functional version where type annotations are required. The next
step, which I won't be doing today, is to add those type annotations to
every function in the compiler, and then use the prestaging version to
compile _that_ version as staging. There may be fits and starts along
the way.