Better transform function; wrapped short circuits in parser
Updated the transform function handle (hopefully) all legal programs.
However, it still duplicates code in conditionals, meaning that I had to
wrap the desugaring of short circuit operators into conditionals in
parse.vzp so that the program could be transformed in a reasonable
amount of memory. I will remove those wrappers when I have a version of
transform that doesn't duplicate code, or when we get rid of short
circuiting.